Fix bad view angles on frozen players.. but at a cost.
From an observer's point of view, the frozen player's pitch angle is
reversed.  e.g. Someone frozen while looking down appears to be severely
looking up to other players.

This is because of the pitch angle (x component) having reverse meaning
between angles and v_angle, and also angles_x being 1/3 of v_angle_x.

This seemed like a less severe bug and worth living with, as the
alternative was having frozen players' viewing angles snap to an
unexpected pitch and taking on a permanent roll.